+++ test/unit/fuzz-tools.test.js
@@ -231,6 +231,7 @@ import * as anthroV238 from '../../lib/anthro-v238.js';
 import * as gisurgV239 from '../../lib/gisurg-v239.js';
 import * as rehabV240 from '../../lib/rehab-v240.js';
 import * as geriV241 from '../../lib/geri-v241.js';
+import * as environV242 from '../../lib/environ-v242.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -406,6 +407,7 @@ const MODULES = {
   'gisurg-v239.js': gisurgV239,
   'rehab-v240.js': rehabV240,
   'geri-v241.js': geriV241,
+  'environ-v242.js': environV242,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];